// src/app/model/data-point.ts
export class DataPoint {
    Id :number;
    Name : string;
    Description : string;
    IsActive : boolean;
    BindingType: string;
    ObjectType : number;
    Connection: string;
    ConnectionDescription: string;
    Device: string;
    DeviceDescription: string;
    Unit: string;
    AlarmConditionsEnabled : boolean;
}